Create new constants file to separate helpers from base
