+++ frontend/src/etc/configuration/events.ts
@@ -10,7 +10,7 @@ import {
   nodeThreatsProperty,
   nodeCommentProperty,
   queueProperty,
-} from "@/etc/constants/base";
+} from "@/etc/constants/common";
 import {
   eventPropertyTypes,
   eventAlertTimeProperty,
